remove trane2, add back flight-delay
