+++ flight-delay/flight-delays/meta.csv
@@ -1,0 +1,20 @@
+name,type,subtype,is_entity
+DATE,time,,FALSE
+DAY_OF_WEEK,categorical,categorical,TRUE
+AIRLINE,categorical,categorical,TRUE
+FLIGHT_NUMBER,id,,TRUE
+TAIL_NUMBER,id,,TRUE
+ORIGIN_AIRPORT,id,,TRUE
+DESTINATION_AIRPORT,id,,TRUE
+SCHEDULED_DEPARTURE_HOUR,number,integer,
+SCHEDULED_TIME,number,float,
+ELAPSED_TIME,number,float,
+DEPARTURE_DELAY,number,float,
+ARRIVAL_DELAY,number,float,
+CANCELLED,categorical,categorical,
+CANCELLATION_REASON,categorical,categorical,
+AIR_SYSTEM_DELAY,number,float,
+SECURITY_DELAY,number,float,
+AIRLINE_DELAY,number,float,
+LATE_AIRCRAFT_DELAY,number,float,
+WEATHER_DELAY,number,float,
