--- trane2/__init__.py
@@ -1,8 +1,0 @@
-from .core import *  # noqa
-from .utils import *  # noqa
-
-import logging
-logname = 'trane.log'
-logging.basicConfig(filename=logname,
-                    filemode='w',
-                    level=logging.INFO)
