--- trane2/utils/__init__.py
@@ -1,3 +1,0 @@
-from .table_meta import *  # noqa
-from .data_parser import *  # noqa
-# from .generate_nl_description import *  # noqa
